requirements.txt
aiohttp==3.11.10
numpy==2.2.0
Pillow==11.0.0
scikit-image
